+++ cmd/titania.go/exec.go
@@ -8,10 +8,17 @@ func exec(directories []string) []*tester.Outcome {
 	outcomes := make([]*tester.Outcome, 0)
 
 	for _, dirname := range directories {
-		tunit := tester.NewTestUnit(dirname)
+		// 設定
+		tconf := tester.NewConfig(dirname)
+		if tconf == nil {
+			continue
+		}
+
+		tunit := tester.NewTestUnit(dirname, tconf)
 		if tunit == nil {
 			continue
 		}
+
 		outcomes = append(outcomes, tunit.Exec())
 	}