+++ cmd/titania.go/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bufio"
 	"encoding/json"
 	"os"
 
@@ -26,7 +27,11 @@ func main() {
 
 	final(outcomes)
 
-	enc := json.NewEncoder(os.Stdout)
+	// buffering
+	stdout := bufio.NewWriter(os.Stdout)
+	defer stdout.Flush()
+
+	enc := json.NewEncoder(stdout)
 	enc.SetIndent("", "  ")
 	enc.SetEscapeHTML(false)