+++ pkg/paizaio/client.go
@@ -2,7 +2,6 @@ package paizaio
 
 import (
 	"bufio"
-	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -63,8 +62,8 @@ func (c *Client) RunnersCreate(req *RunnersCreateRequest) (*RunnersCreateRespons
 		return res, err
 	}
 
-	if res.ID == "" {
-		return res, errors.New(res.Error)
+	if res.Error != "" {
+		return res, &RunnerError{res.Error}
 	}
 
 	return res, nil
@@ -76,14 +75,21 @@ func (c *Client) RunnersGetStatus(req *RunnersGetStatusRequest) (*RunnersGetStat
 		return res, err
 	}
 
-	return res, nil
+	if res.Error != "" {
+		return res, &RunnerError{res.Error}
+	}
 
+	return res, nil
 }
 
 func (c *Client) RunnersGetDetails(req *RunnersGetDetailsRequest) (*RunnersGetDetailsResponse, error) {
 	res := &RunnersGetDetailsResponse{}
 	if err := c.api("GET", "/runners/get_details", req, res); err != nil {
 		return res, err
+	}
+
+	if res.Error != "" {
+		return res, &RunnerError{res.Error}
 	}
 
 	return res, nil