+++ pkg/paizaio/errors.go
@@ -10,10 +10,18 @@ type ServerError struct {
 	msg  string
 }
 
+type RunnerError struct {
+	msg string
+}
+
 func (e ClientError) Error() string {
 	return e.msg
 }
 
 func (e ServerError) Error() string {
+	return e.msg
+}
+
+func (e RunnerError) Error() string {
 	return e.msg
 }